requirements.txt
cohere
langchain
streamlit
python-dotenv
PyPDF2 


-e .